web search working fine now
